+++ trunk/src/plugins/select/bluegene/plugin/bluegene.h
@@ -132,7 +132,7 @@ extern int init_bg(void);
 /* Purge all plugin variables */
 extern void fini_bg(void);
 
-extern char *give_geo(int int_geo[SYSTEM_DIMENSIONS]);
+extern char *give_geo(uint16_t int_geo[SYSTEM_DIMENSIONS]);
 
 extern bool blocks_overlap(bg_record_t *rec_a, bg_record_t *rec_b);
